requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
dj-database-url==1.0.0
Django==3.2.5
django-admin-interface==0.19.1
django-bootstrap4==22.1
django-ckeditor==6.3.2
django-colorfield==0.7.0
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
whitenoise==6.2.0
